Extract ci-ready gate predicates under anti-monolith ceiling (#2502)
The new is_ci_ready_terminal_at_head predicate plus its tests pushed
src/daemon/pr_state/mod.rs past its grandfathered LOC ceiling (3428 →
3513), tripping tests/src_file_size_invariant.rs in CI. Move BOTH
ci-ready emit-gate predicates (is_ci_ready_merge_blocked +
is_ci_ready_terminal_at_head) and their unit tests into a focused
src/daemon/pr_state/ready_gate.rs, re-exported from the parent module so
every call site (crate::daemon::pr_state::is_ci_ready_*,
super::is_ci_ready_merge_blocked) resolves unchanged. mod.rs drops to
3370 LOC; behavior is byte-for-byte identical.

Claim: src/daemon/pr_state/ready_gate.rs holds both gate predicates + tests, built via new_for_branch
Claim: src/daemon/pr_state/mod.rs declares pub mod ready_gate + re-exports both predicates, removing them and their tests inline
