+++ src/daemon/pr_state/mod.rs
@@ -49,8 +49,14 @@ use serde::{Deserialize, Serialize};
 
 pub mod auto_arm;
 pub mod gh_poll;
+pub mod ready_gate;
 mod remote_gc;
 mod scanner;
+
+// #2502: ci-ready emit-gate predicates live in `ready_gate` (extracted to keep
+// this file under its anti-monolith ceiling); re-exported so existing call sites
+// resolve unchanged.
+pub use ready_gate::{is_ci_ready_merge_blocked, is_ci_ready_terminal_at_head};
 pub(crate) mod verdict_buffer;
 // #986: the production per-tick handler drives the scanner with an explicit
 // (snapshot) poller via `scan_and_emit_with` — the old `scan_and_emit` wrapper
@@ -444,52 +450,6 @@ pub fn is_merge_ready(state: &PrState) -> bool {
     reviewers
         .iter()
         .all(|(_, reviewed)| sha_prefix_match(&state.head_sha, reviewed))
-}
-
-/// #t-92758: a PR whose `[ci-ready-for-action]` chain handoff is pointless right
-/// now because the PR cannot be merged/acted-on — a REJECTED verdict (a reviewer
-/// bounced it; it's being reworked) or a Draft PR (`gh pr merge` refuses drafts).
-/// Used to (a) SUPPRESS a new ci-ready emission and (b) EVICT an existing
-/// ci-handoff track so the re-nudge watchdog stops pinging the chain target for a
-/// PR they can't move.
-///
-/// IRON RULE (regression-pinned): this is DELIBERATELY narrow — it returns
-/// `false` for `Verified` / `Unverified` / `Pending` / `None` verdicts. A
-/// VERIFIED+green PR is exactly the "your turn to merge" case the chain exists
-/// for and MUST keep emitting + re-nudging; this predicate must never suppress
-/// it. (Unverified/Pending/None are not merge-blocked verdicts — the reviewer
-/// hasn't bounced the PR — so the chain handoff stays live.)
-pub fn is_ci_ready_merge_blocked(state: &PrState) -> bool {
-    matches!(state.verdict_state, VerdictState::Rejected { .. })
-        || matches!(state.draft_state, DraftState::Draft)
-}
-
-/// #2502: a PR whose cached terminal merge_state (Merged / ClosedUnmerged) was
-/// observed at the SAME head CI just passed on. Emitting `[ci-ready-for-action]`
-/// here hands "your turn" on a PR that is already merged/closed, spawning a
-/// re-nudge loop the chain target can't resolve — the proactive emit-time half of
-/// a suppression whose reactive half is the scanner's terminal track-evict
-/// (`pr_state::scanner`, the green-then-terminal ordering).
-///
-/// HEAD-GUARDED (load-bearing): a terminal state at a DIFFERENT head — a
-/// force-push / branch-reuse that opened a fresh PR on the same branch — returns
-/// `false`, so the live handoff still fires. This is sound because
-/// [`record_ci_result`] SKIPS `CiObserved` on terminal states (#1314), freezing a
-/// terminal file's `head_sha` at the merge/close head; `state.head_sha == ci_head`
-/// is therefore a reliable proxy for "CI ran on the head that was merged/closed",
-/// never a later reused head.
-///
-/// Distinct from [`is_ci_ready_merge_blocked`] (REJECTED/Draft, still-OPEN): that
-/// predicate is IRON-RULE-narrow and MUST NOT be widened to terminal states. The
-/// emit site ORs the two. Both fail OPEN — a missing sidecar (caller passes
-/// `None`), a non-terminal merge_state, an empty `head_sha`, or a head mismatch
-/// all return `false` and the normal emit proceeds.
-pub fn is_ci_ready_terminal_at_head(state: &PrState, ci_head: &str) -> bool {
-    matches!(
-        state.merge_state,
-        MergeState::Merged { .. } | MergeState::ClosedUnmerged { .. }
-    ) && !state.head_sha.is_empty()
-        && state.head_sha == ci_head
 }
 
 // ─── storage ───────────────────────────────────────────────────────────
@@ -1297,109 +1257,6 @@ mod tests {
             created_at: now(),
             updated_at: now(),
         }
-    }
-
-    /// #t-92758 IRON RULE: `is_ci_ready_merge_blocked` blocks ONLY a REJECTED
-    /// verdict or a Draft PR — never VERIFIED / Unverified / Pending / None. A
-    /// VERIFIED+green PR is the "your turn to merge" case the ci-ready chain
-    /// exists for and MUST keep emitting + re-nudging; a regression that made the
-    /// predicate true for VERIFIED would silently kill legitimate merge handoffs.
-    #[test]
-    fn is_ci_ready_merge_blocked_only_rejected_or_draft() {
-        let mut s = new_state("sha-A", ReviewClass::Single);
-
-        // Non-blocking verdicts (Ready draft state):
-        s.verdict_state = VerdictState::None;
-        assert!(!is_ci_ready_merge_blocked(&s), "None must not block");
-        s.verdict_state = VerdictState::Pending;
-        assert!(!is_ci_ready_merge_blocked(&s), "Pending must not block");
-        s.verdict_state = VerdictState::Unverified {
-            reviewer: "r".into(),
-            reviewed_head: "sha-A".into(),
-        };
-        assert!(!is_ci_ready_merge_blocked(&s), "Unverified must not block");
-        s.verdict_state = VerdictState::Verified {
-            reviewers: vec![("r".into(), "sha-A".into())],
-        };
-        assert!(
-            !is_ci_ready_merge_blocked(&s),
-            "IRON RULE: VERIFIED must NEVER be suppressed/evicted"
-        );
-
-        // Blocking: REJECTED verdict.
-        s.verdict_state = VerdictState::Rejected {
-            reviewer: "r".into(),
-            reviewed_head: "sha-A".into(),
-            reason: None,
-        };
-        assert!(is_ci_ready_merge_blocked(&s), "REJECTED must block");
-
-        // Blocking: Draft — even with an otherwise-mergeable VERIFIED verdict.
-        s.verdict_state = VerdictState::Verified {
-            reviewers: vec![("r".into(), "sha-A".into())],
-        };
-        s.draft_state = DraftState::Draft;
-        assert!(
-            is_ci_ready_merge_blocked(&s),
-            "Draft must block even with a VERIFIED verdict"
-        );
-    }
-
-    /// #2502: `is_ci_ready_terminal_at_head` suppresses ONLY a terminal
-    /// (Merged / ClosedUnmerged) PR observed at the SAME head CI passed on, and
-    /// is head-GUARDED — a terminal state at a different head (force-push /
-    /// branch-reuse) MUST NOT suppress, or a fresh PR on a reused branch would
-    /// silently lose its handoff. Non-terminal states never suppress.
-    #[test]
-    fn is_ci_ready_terminal_at_head_only_same_head_terminal() {
-        let mut s = new_state("sha-A", ReviewClass::Single);
-
-        // Non-terminal states never suppress, even at the matching head.
-        s.merge_state = MergeState::NotReady;
-        assert!(
-            !is_ci_ready_terminal_at_head(&s, "sha-A"),
-            "NotReady must not suppress"
-        );
-        s.merge_state = MergeState::MergeReady;
-        assert!(
-            !is_ci_ready_terminal_at_head(&s, "sha-A"),
-            "MergeReady must not suppress"
-        );
-
-        // Merged at the SAME head → suppress.
-        s.merge_state = MergeState::Merged {
-            merge_commit: "mc".into(),
-            merged_at: now(),
-        };
-        assert!(
-            is_ci_ready_terminal_at_head(&s, "sha-A"),
-            "Merged at same head must suppress"
-        );
-        // Merged at a DIFFERENT head (force-push / reuse) → fail open. This is the
-        // anti-false-suppression nail: #1314 freezes a terminal head_sha at the
-        // merge head, so a green at sha-B is a fresh PR on the reused branch.
-        assert!(
-            !is_ci_ready_terminal_at_head(&s, "sha-B"),
-            "Merged at a different head must NOT suppress (branch reuse stays live)"
-        );
-
-        // ClosedUnmerged at the SAME head → suppress; different head → fail open.
-        s.merge_state = MergeState::ClosedUnmerged { closed_at: now() };
-        assert!(
-            is_ci_ready_terminal_at_head(&s, "sha-A"),
-            "ClosedUnmerged at same head must suppress"
-        );
-        assert!(
-            !is_ci_ready_terminal_at_head(&s, "sha-B"),
-            "ClosedUnmerged at a different head must NOT suppress"
-        );
-
-        // Empty head_sha → fail open (never match an empty ci_head either).
-        s.head_sha = String::new();
-        assert!(
-            !is_ci_ready_terminal_at_head(&s, ""),
-            "empty head_sha must fail open (no spurious suppression)"
-        );
     }
 
     /// T1: CI green at head_sha + Verified at same head_sha → MergeReady.
